pkair

+++ plonky2x/src/frontend/hash/blake2/blake2b_curta.rs
@@ -111,7 +111,7 @@ impl<L: PlonkParameters<D>, const D: usize> CircuitBuilder<L, D> {
             padded_messages.extend(curta_req.message.clone());
             msg_lengths.push(curta_req.message_len);
             digests.extend(curta_req.digest);
-            chunk_sizes.push(curta_req.chunk_size);
+            chunk_sizes.push(curta_req.chunk_size as u64);
         }
 
         let mut blake2b_builder_gadget: BLAKE2BBuilderGadget<
